Parse new fields in the front
