+++ front/packages/ui/src/details/show.tsx
@@ -85,7 +85,7 @@ const query = (slug: string): QueryIdentifier<Show> => ({
 	parser: ShowP,
 	path: ["shows", slug],
 	params: {
-		fields: ["studio"],
+		fields: ["studio", "firstEpisode"],
 	},
 });
